Remove smart proxy routes from foreman_puppet namespace
